+++ src/api/git-current-branch.js
@@ -1,0 +1,7 @@
+import { commander } from '../commander.js'
+
+export async function currentBranch () {
+  const rawText = await commander('git rev-parse --abbrev-ref HEAD')
+
+  return rawText
+}