+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+- Added function `currentBranch()` to the API to get the current branch name.
+
 ## [1.0.4] 2023-02-26
 
 ### What's Changed